pyproject.toml
[project]
name = "banking-snowflake-dbt-airflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow>=3.1.7",
    "boto3>=1.42.59",
    "confluent-kafka>=2.13.0",
    "dbt-snowflake>=1.11.2",
    "faker>=40.5.1",
    "fastparquet>=2025.12.0",
    "pandas>=3.0.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]
